cql3: select_statement: disambiguate execute() overloads
There are two execute() overloads, but they don't do the same thing - one is a
partial implementation of the other.

The same is true of two execute_without_checking_exception_message() overloads.

Change the name of the subordinate overload to indicate its role. Overloads should
be used when the only difference between overloads is the argument type, not when
one does a subset of the other.