types: add read_nth_tuple_element()
Add a function which retrieves the value of nth
field from a serialized tuple value.

I tried to make it as efficient as possible.
Other functions, like evaluate(subscript) tend to
deserialize the whole structure and put all of its
elements in a vector. Then they select a single element
from this vector.
This is wasteful, as we only need a single element's value.

This function goes over the serialized fields
and directly returns the one that is needed.
No allocations are needed.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>