repair: set repair state correctly
repair_node_state::state is only for debugging purpose, see
ab57cea7839ce813d96acf233edf905fc88c2795 which introduced it.
so this change does not impact the behavior of scylla, but can
improve the debugging experience by reflecting more accurate state
of repair when we are actually inspecting it.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>