column_identifer: add column_identifier_raw::text()
I would like to be able to get a reference to the
string inside `column_identifer_raw`, but there was
no such function. There was only `to_string()`, which
copies the entire string, which is wasteful.

Let's add the method `text()`, which returns a reference
instead of a copy. `column_identifier` already has
such method, so `column_identifier_raw` can have one
as well.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>