test/alternator: add process ID to unique table name
To allow multiple tests to be run on the same DynamoDB API instance
(AWS or an invocation of ScyllaDB Alternator), each tables we create
for the tests has a unique name. Before this patch, the unique name
was based on the current time. This was usually good enough, but if
multiple test runs are run in parallel, the two can happen to pick
the same "unique" table name and fail.

So in this patch we add to the unique table name also the process id
of the test. There is still a theoretical possibility of two tests
running at the same time on two machines and happening to get the
same process id by chance, but the chance of this happening is
absurdly small (on modern Linux installations, process ids are chosen
in the range up to 4 million).

One test that assumed the previous *length* of the table name had to
be modifed to pass after we changed the table name.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>