test/alternator: add "-j N" option to run tests in parallel
Before this patch, test/alternator/run runs Scylla and then runs a
single invocation of "pytest" to run all tests. Pytest runs the
different test functions (we have 650 of those by now) in sequence,
one after another, with no concurrency.

This patch adds to test/alternator/run the option "-j N" (e.g., "-j 6").
With this option, instead of running a single pytest invocation we run
pytest on each test file separately (today we have 33 of those), and
moreover run up to N of those in parallel.

The "-j N" option can be used to to finish the tests faster (on my
laptop on a dev build, "run" finishes in 25 seconds, "run -j 6"
finishes in 15 seconds, as well as to ensure that the Scylla process
uses up all the CPU cores it was allocated and not less.

The "-j 1" option isn't the same as no -j option - it runs pytest
on each source file separately, and is slower than the default,
so it's not recommended.

Note that even after this patch, test/alternator/run runs Scylla on
two shards by default. This number is not yet configurable.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>