sstable: do not rename if tempdir is not specified
before this change, we already rename the component file if it is
opened for write, even if the temp_dir is not set. in that case,
the underlying rename() syscall is a noop, but it still a waste.

so, instead of checking `readonly`, we just check if the `tgt_dir`
is the sstable's dir. this is not only more readable, but also more
correct in the sense that it avoids unnecessary syscall.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>