gossip: Do not advertise SHUTDOWN gossip status
The SHUTDOWN status is a special NORMAL status. A node in SHUTDOWN status is
part of the ring the same as NORMAL status. To simplify the handling, we
can get rid of the SHUTDOWN status completely. There is no real
difference between a node in SHUTDOWN status and a in NORMAL + DOWN
status.

After this patch, when a node shutdown:

- The shutdown node does not advertise SHUTDOWN gossip status
- The shutdown node still sends the the SHUTDOWN rpc message
- The receivers of the SHUTDOWN rpc message do not change the status of the
  node to SHUTDOWN but only mark the node as DOWN, keep the node as
  NORMAL gossip status.

This way, other nodes still can mark the shutdown node as DOWN
immediately while we have less gossip status in the system to handle and
reason about.

For example, with this patch, we can now get rid of the hacks in
f3bd7e9b8c6 to add LEFT status to avoid sending shutdown message.

```
// Need to set it otherwise gossiper will try to send shutdown on exit
co_await _gossiper.add_local_application_state({{
    gms::application_state::STATUS, gms::versioned_value::left({},
    _gossiper.now().time_since_epoch().count()) }});
```